Updated inventory system
Getting close to finishing it, added the ability to have a maximum number of items in a slot and what to do with leftover items. It's a small bug where it just fills all slots now no matter what. Just need to check the code a little later and it should be all sorted.